# docs/api.yaml
openapi: 3.0.3
info:
  title: Flask MongoDB API
  description: REST API for managing customers and subscriptions
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /api/customers:
    get:
      summary: Get all customers
      description: Retrieve a list of customers with optional filtering and pagination
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: status
          in: query
          description: Filter by customer status
          schema:
            type: string
            enum: [active, inactive, suspended]
        - name: search
          in: query
          description: Search customers by name or email
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: object
                properties:
                  customers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Customer"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Create a new customer
      description: Create a new customer with the provided data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerCreate"
      responses:
        "201":
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Customer with email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/customers/{customer_id}:
    get:
      summary: Get a customer by ID
      description: Retrieve a specific customer by their ID
      parameters:
        - name: customer_id
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Invalid customer ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update a customer
      description: Update an existing customer's information
      parameters:
        - name: customer_id
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Validation error or invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete a customer
      description: Delete a customer (only if they have no active subscriptions)
      parameters:
        - name: customer_id
          in: path
          required: true
          description: Customer ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Customer deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Customer deleted successfully
        "400":
          description: Invalid ID or customer has active subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Customer:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        phone:
          type: string
          example: "+1234567890"
        address:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        country:
          type: string
          example: "USA"
        postal_code:
          type: string
          example: "10001"
        status:
          type: string
          enum: [active, inactive, suspended]
          example: "active"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CustomerCreate:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        phone:
          type: string
          maxLength: 20
          example: "+1234567890"
        address:
          type: string
          maxLength: 200
          example: "123 Main St"
        city:
          type: string
          maxLength: 100
          example: "New York"
        country:
          type: string
          maxLength: 100
          example: "USA"
        postal_code:
          type: string
          maxLength: 20
          example: "10001"
        status:
          type: string
          enum: [active, inactive, suspended]
          default: active
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        phone:
          type: string
          maxLength: 20
          example: "+1234567890"
        address:
          type: string
          maxLength: 200
          example: "123 Main St"
        city:
          type: string
          maxLength: 100
          example: "New York"
        country:
          type: string
          maxLength: 100
          example: "USA"
        postal_code:
          type: string
          maxLength: 20
          example: "10001"
        status:
          type: string
          enum: [active, inactive, suspended]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 10
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Resource not found"
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation error"
        details:
          type: object
          example: { "email": ["Invalid email format"] }
...